chore: click directly on vulnerability to get modal and change button of repositories
