+++ src/components/Button.jsx
@@ -38,11 +38,11 @@ const Button = ({ url, languages }) => {
 
         <input
           type="submit"
-          value={`${isActive ? "Scan" : "Unavailable"}`}
+          value={`${isActive ? "Scan now!" : "Unavailable"}`}
           disabled={!isActive}
           className={`${
             isActive
-              ? "cursor-pointer bg-gradient-to-r from-purple-400 to-pink-600 font-bold p-2 rounded-md hover:opacity-60 transition ease-in-out delay-50 text-center w-full md:w-fit"
+              ? "cursor-pointer hover:scale-90 ml-2 px-2 text-[10px] bg-gradient-to-r from-purple-400 to-pink-600 font-bold rounded-sm hover:opacity-60 transition ease-in-out delay-50"
               : "text-gray-600 border-gray-600 w-fit rounded-md"
           }
