+++ src/components/Ecosystem.jsx
@@ -22,7 +22,9 @@ const Ecosystem = ({
       onClick={handleClick}
       className={`flex flex-row items-center gap-x-5 p-2  ${
         status === "unavailable" ? "text-gray-600" : null
-      } rounded-sm mr-3 ${isActive ? "bg-gray-600" : null}`}
+      } rounded-sm mr-3 hover:scale-90 transition ease-in-out duration-300 ${
+        isActive ? "bg-gray-600 " : null
+      }`}
     >
       {name === "ERLANG" ? "ELIXIR" : name}
       {loading ? <Loader /> : <Status status={status} />}
