+++ src/components/SocialNetwork.jsx
@@ -7,14 +7,14 @@ const SocialNetwork = () => {
       <a
         href="https://github.com/francois973/"
         target="_blank"
-        className="hover:scale-110 transition ease-out"
+        className="hover:scale-90 transition ease-out"
       >
         <AiFillGithub className="w-10 h-10" />
       </a>
       <a
         href="https://www.linkedin.com/in/francoisrobino/"
         target="_blank"
-        className="hover:scale-110 transition ease-out"
+        className="hover:scale-90 transition ease-out"
       >
         <AiFillLinkedin className="w-10 h-10" />
       </a>
