+++ src/components/StartForm.jsx
@@ -49,7 +49,7 @@ const StartForm = () => {
         <input
           type="submit"
           value="Scan now!"
-          className="cursor-pointer ml-2 px-2 text-[10px] bg-gradient-to-r from-purple-400 to-pink-600 font-bold rounded-sm hover:opacity-60 transition ease-in-out delay-50"
+          className="cursor-pointer hover:scale-90 ml-2 px-2 text-[10px] bg-gradient-to-r from-purple-400 to-pink-600 font-bold rounded-sm hover:opacity-60 transition ease-in-out delay-50"
         />
       </div>
       <span className="text-[9px] text-gray-500">
