+++ src/components/Vulnerabilities.jsx
@@ -10,7 +10,7 @@ const Vulnerabilities = ({ vulnerabilities, selectedEcosystem }) => {
 
   return (
     <div>
-      <ul className="flex flex-wrap gap-y-8 justify-center gap-x-3 lg:flex-row">
+      <ul className="flex flex-wrap gap-y-8 justify-center gap-x-3 lg:flex-row pb-20">
         {filteredVulnerabilities.length === 0 ? (
           <p>No vulnerabilities here !</p>
         ) : (
