+++ src/pages/Project.jsx
@@ -51,19 +51,19 @@ function Project() {
   return (
     <div>
       <div className="ml-7 mt-20">
-        <h2 className="text-lg mb-2 ">
+        <h2 className="text-lg mb-2 text-[16px]">
           <strong>Repository:</strong>{" "}
           <p className="sm:text-[16px] text-[12px] pr-20">{project.url}</p>
         </h2>
-        <h3 className="mb-2">
+        <h3 className="mb-2 text-[16px]">
           <strong>Created at:</strong>{" "}
           {project.created_at.split("T")[1].split(".")[0]}
         </h3>
-        <p>
+        <p className="text-[16px]">
           <strong>Vulnerabilities detected:</strong> {vulnerabilities.length}
         </p>
       </div>
-      <div className="ml-3 mb-7">
+      <div className="ml-3 mb-7 text-[16px]">
         <Ecosystems
           ecosystems={ecosystems}
           setSelectedEcosystem={setSelectedEcosystem}
